secret_chat_client_tkinter/requirements.txt
pytz==2019.2
aiofiles==0.4.0
python-dotenv==0.10.3
